add validate /delete function

+++ api/web/api/post/views.py
@@ -41,7 +41,13 @@ async def delete_post(
     postid: PostID,
     post_dao: PostDAO = Depends(),
 ):
-    
+    post_data = await post_dao.get_post_by_id(postid)
+
+    if post_data is None:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Post not found. Please check the post ID and try again."
+        )
     
     return await post_dao.delete_post(
         postid=postid.postid
